authentication_service/requirements.txt
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.5
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.22.1
passlib==1.7.4
pycparser==2.20
pymongo==3.11.2
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
speaklater==1.3
Werkzeug==1.0.1
WTForms==2.3.3
